Move ApiState to its context provider
